Moved formatTo implementation into MessageChannel rather than sub-interfaces
